WIP Rely on Whitehall::Authority for publishing perms
As with access limiting, the rules that govern whether a user can
publish or force publish an edition are (mostly) present and correct
in Whitehall::Authority. This removes any knowledge of the users
publishing permissions from the Edition model as we can rely on the
controller to guard based on permissions.

The one exception was the permission on publishing a scheduled
edition. This is what guards that only the scheduled publishing robot
can publish scheduled editions. I've updated the EditionRules to
reflect this.

NOTE: There is work remaining here and that is to prise apart
publishing and force publishing so that they are separate transitions
rather than force publishing being an option that is passed to the
publish_as method. This is so that we can have separate
Edition#can_publish? and  Edition#can_force_publish? to verify that
the edition is in an appropriate state to perform the action